13287 - Separate service and package puppet code into util classes
In the past the service and package agent have been totally reliant on
puppet being installed.

The puppet implementation have been seperated from the service and
package agents. The puppet-package and puppet-service agents have
been renamed to package and service respectively.

The actual service and package implementations are now found in
util/puppetpackage.rb and util/puppetservice.rb and can be interchanged
with other implementations. The one that is going to be used in specified
in server.cfg for example

plugin.package.provider = puppet

  and

plugin.service.provider = puppet