17400 - ActionPolicy doesn't properly handle multiple facts/classes in policy directive
In the past all Fact and Class lists would apply a logical 'and' to all
the given values, limiting the flexibility of the policy.

Fact and Class fields can now be defined using the mcollective discovery
language inside the policy file. Facts and Classes can be combined by
dropping the class field in the policy file and listing both facts and
classes as a complex statement in the facts field.

The Rakefile has also been updated to include simplerpc_authorization tests.